Redo theme fetching
