+++ src/utils/index.js
@@ -1,17 +1,13 @@
 import slugify from './slugify'
 import splitString from './splitString'
 import stringToDot from './stringToDot';
-import getThemeClasses from './getThemeClasses';
-import getVariantClasses from './getVariantClasses';
-import getStatusClasses from './getStatusClasses';
 import filterClasses from './filterClasses';
+import themer from './themer';
 
 export {
+  themer,
   slugify,
   splitString,
   stringToDot,
-  getThemeClasses,
-  getVariantClasses,
-  getStatusClasses,
   filterClasses
 }
